# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: SuqCheck API
  description: Evidence-backed price estimates for packaged household essentials in
    Ethiopia.
  version: 0.1.0
paths:
...
  /api/products:
    get:
      tags:
      - products
      summary: List Products
      operationId: list_products_api_products_get
      parameters:
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Q
      - name: category
        in: query
        required: false
        schema:
          anyOf:
          - enum:
            - cooking_oil
            - sugar
            - rice
            - flour
            - salt
            - pasta
            - coffee
            - tea
            - milk
            - soap
            - detergent
            - toothpaste
            - shampoo
            - bottled_water
            type: string
          - type: 'null'
          title: Category
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/products/{product_id}:
    get:
      tags:
      - products
      summary: Get Product
      operationId: get_product_api_products__product_id__get
      parameters:
      - name: product_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Product Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDetail'
        '404':
          description: No product with this id has a priced estimate.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/products/{product_id}/stores:
    get:
      tags:
      - products
      summary: List Product Stores
      operationId: list_product_stores_api_products__product_id__stores_get
      parameters:
      - name: product_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Product Id
      - name: latitude
        in: query
        required: false
        schema:
          anyOf:
          - type: number
            maximum: 90
            minimum: -90
          - type: 'null'
          title: Latitude
      - name: longitude
        in: query
        required: false
        schema:
          anyOf:
          - type: number
            maximum: 180
            minimum: -180
          - type: 'null'
          title: Longitude
      - name: radius_m
        in: query
        required: false
        schema:
          type: integer
          maximum: 50000
          minimum: 100
          default: 5000
          title: Radius M
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NearbyStoresResponse'
        '404':
          description: No product with this id has a priced estimate.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/stores/{store_id}:
    get:
      tags:
      - stores
      summary: Get Store
      operationId: get_store_api_stores__store_id__get
      parameters:
      - name: store_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Store Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreDetail'
        '404':
          description: No store with this id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/evidence:
    get:
      tags:
      - evidence
      summary: List Evidence
      operationId: list_evidence_api_evidence_get
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - enum:
            - accepted
            - pending
            - rejected
            type: string
          - type: 'null'
          description: Filter by the gate decision.
          title: Status
        description: Filter by the gate decision.
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvidenceLogResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/evidence/receipt:
    post:
      tags:
      - evidence
      summary: Upload Receipt
      operationId: upload_receipt_api_evidence_receipt_post
      parameters:
      - name: X-Device-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Anonymous device identifier used for rate limiting.
          title: X-Device-Id
        description: Anonymous device identifier used for rate limiting.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_receipt_api_evidence_receipt_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptUploadResponse'
        '429':
          description: The per-device or per-network upload limit was reached. The
            `Retry-After` header carries the wait in seconds.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: The image is larger than 8MB.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '415':
          description: The image must be one of image/jpeg, image/png, image/webp.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: The image could not be read; ask the shopper to retake it.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/evidence/shelf:
    post:
      tags:
      - evidence
      summary: Upload Shelf Photo
      operationId: upload_shelf_photo_api_evidence_shelf_post
      parameters:
      - name: X-Device-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Anonymous device identifier used for rate limiting.
          title: X-Device-Id
        description: Anonymous device identifier used for rate limiting.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_shelf_photo_api_evidence_shelf_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShelfUploadResponse'
        '429':
          description: The per-device or per-network upload limit was reached. The
            `Retry-After` header carries the wait in seconds.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: The image is larger than 8MB.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '415':
          description: The image must be one of image/jpeg, image/png, image/webp.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: The image could not be read; ask the shopper to retake it.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: The tagged product is not in the catalog, so its price cannot
            be recorded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/evidence/manual:
    post:
      tags:
      - evidence
      summary: Submit Manual Evidence
      operationId: submit_manual_evidence_api_evidence_manual_post
      parameters:
      - name: X-Device-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Anonymous device identifier used for rate limiting.
          title: X-Device-Id
        description: Anonymous device identifier used for rate limiting.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManualEvidenceRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManualEvidenceResponse'
        '429':
          description: The per-device or per-network upload limit was reached. The
            `Retry-After` header carries the wait in seconds.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: The report names a product or store that does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/scan/identify:
    post:
      tags:
      - scan
      summary: Identify Product
      operationId: identify_product_api_scan_identify_post
      parameters:
      - name: X-Device-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Anonymous device identifier used for rate limiting.
          title: X-Device-Id
        description: Anonymous device identifier used for rate limiting.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_identify_product_api_scan_identify_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductIdentification'
        '429':
          description: The per-device or per-network upload limit was reached. The
            `Retry-After` header carries the wait in seconds.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: The image is larger than 8MB.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '415':
          description: The image must be one of image/jpeg, image/png, image/webp.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: The image could not be read; ask the shopper to retake it.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/analytics/trends:
    get:
      tags:
      - analytics
      summary: Get Trends
      operationId: get_trends_api_analytics_trends_get
      parameters:
      - name: period_days
        in: query
        required: false
        schema:
          type: integer
          maximum: 90
          minimum: 2
          default: 7
          title: Period Days
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrendsResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    Body_identify_product_api_scan_identify_post:
      properties:
        image:
          type: string
          contentMediaType: application/octet-stream
          title: Image
          description: JPEG, PNG, or WebP product image.
      type: object
      required:
      - image
      title: Body_identify_product_api_scan_identify_post
    Body_upload_receipt_api_evidence_receipt_post:
      properties:
        image:
          type: string
          contentMediaType: application/octet-stream
          title: Image
          description: JPEG, PNG, or WebP receipt image.
      type: object
      required:
      - image
      title: Body_upload_receipt_api_evidence_receipt_post
    Body_upload_shelf_photo_api_evidence_shelf_post:
      properties:
        image:
          type: string
          contentMediaType: application/octet-stream
          title: Image
          description: JPEG, PNG, or WebP shelf photo.
      type: object
      required:
      - image
      title: Body_upload_shelf_photo_api_evidence_shelf_post
...
    ErrorResponse:
      properties:
        detail:
          type: string
          title: Detail
      type: object
      required:
      - detail
      title: ErrorResponse
      description: Every error this API returns, including 429, has this body.
...
    EvidenceLogResponse:
      properties:
        total:
          type: integer
          minimum: 0.0
          title: Total
        limit:
          type: integer
          minimum: 1.0
          title: Limit
        offset:
          type: integer
          minimum: 0.0
          title: Offset
        items:
          items:
            $ref: '#/components/schemas/EvidenceLogItem'
          type: array
          title: Items
      type: object
      required:
      - total
      - limit
      - offset
      - items
      title: EvidenceLogResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ManualEvidenceRequest:
      properties:
        product_id:
          type: string
          format: uuid
          title: Product Id
        store_id:
          type: string
          format: uuid
          title: Store Id
        price_etb:
          type: number
          exclusiveMinimum: 0.0
          title: Price Etb
        observed_at:
          type: string
          format: date-time
          title: Observed At
        source_type:
          type: string
          enum:
          - store_visit
          - community
          title: Source Type
          default: community
      type: object
      required:
      - product_id
      - store_id
      - price_etb
      - observed_at
      title: ManualEvidenceRequest
    ManualEvidenceResponse:
      properties:
        decision:
          $ref: '#/components/schemas/EvidenceDecision'
      type: object
      required:
      - decision
      title: ManualEvidenceResponse
...
    NearbyStoresResponse:
      properties:
        product_id:
          type: string
          format: uuid
          title: Product Id
        market_price_etb:
          type: number
          exclusiveMinimum: 0.0
          title: Market Price Etb
        items:
          items:
            $ref: '#/components/schemas/NearbyStorePrice'
          type: array
          title: Items
      type: object
      required:
      - product_id
      - market_price_etb
      - items
      title: NearbyStoresResponse
    ProductDetail:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        canonical_name:
          type: string
          title: Canonical Name
        brand:
          type: string
          title: Brand
        category:
          type: string
          enum:
          - cooking_oil
          - sugar
          - rice
          - flour
          - salt
          - pasta
          - coffee
          - tea
          - milk
          - soap
          - detergent
          - toothpaste
          - shampoo
          - bottled_water
          title: Category
        size_label:
          type: string
          title: Size Label
        market_price_etb:
          type: number
          exclusiveMinimum: 0.0
          title: Market Price Etb
        confidence:
          type: integer
          maximum: 100.0
          minimum: 0.0
          title: Confidence
        confidence_band:
          type: string
          enum:
          - high
          - medium
          - low
          title: Confidence Band
        thumbnail_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Thumbnail Url
        barcode:
          anyOf:
          - type: string
          - type: 'null'
          title: Barcode
        price_range_etb:
          prefixItems:
          - type: number
          - type: number
          type: array
          maxItems: 2
          minItems: 2
          title: Price Range Etb
        evidence_count:
          type: integer
          minimum: 0.0
          title: Evidence Count
        store_count:
          type: integer
          minimum: 0.0
          title: Store Count
        spread_pct:
          type: number
          minimum: 0.0
          title: Spread Pct
        updated_at:
          type: string
          format: date-time
          title: Updated At
        confidence_breakdown:
          $ref: '#/components/schemas/ConfidenceBreakdown'
        sources:
          items:
            $ref: '#/components/schemas/SourceSummary'
          type: array
          title: Sources
        history:
          items:
            $ref: '#/components/schemas/HistoryPoint'
          type: array
          title: History
      type: object
      required:
      - id
      - canonical_name
      - brand
      - category
      - size_label
      - market_price_etb
      - confidence
      - confidence_band
      - price_range_etb
      - evidence_count
      - store_count
      - spread_pct
      - updated_at
      - confidence_breakdown
      - sources
      - history
      title: ProductDetail
...
    ProductListResponse:
      properties:
        total:
          type: integer
          minimum: 0.0
          title: Total
        limit:
          type: integer
          minimum: 1.0
          title: Limit
        offset:
          type: integer
          minimum: 0.0
          title: Offset
        items:
          items:
            $ref: '#/components/schemas/ProductSummary'
          type: array
          title: Items
      type: object
      required:
      - total
      - limit
      - offset
      - items
      title: ProductListResponse
...
    ReceiptUploadResponse:
      properties:
        extraction:
          $ref: '#/components/schemas/ReceiptExtraction'
        decisions:
          items:
            $ref: '#/components/schemas/EvidenceDecision'
          type: array
          title: Decisions
      type: object
      required:
      - extraction
      - decisions
      title: ReceiptUploadResponse
...
    ShelfUploadResponse:
      properties:
        extraction:
          $ref: '#/components/schemas/ShelfExtraction'
        decision:
          $ref: '#/components/schemas/EvidenceDecision'
      type: object
      required:
      - extraction
      - decision
      title: ShelfUploadResponse
...
    StoreDetail:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        district:
          type: string
          title: District
        kind:
          type: string
          enum:
          - supermarket
          - shop
          - online
          title: Kind
        latitude:
          type: number
          title: Latitude
        longitude:
          type: number
          title: Longitude
        product_count:
          type: integer
          minimum: 0.0
          title: Product Count
        average_price_index:
          type: number
          title: Average Price Index
          description: 100 is market average; below 100 is cheaper.
        last_reported_at:
          type: string
          format: date-time
          title: Last Reported At
      type: object
      required:
      - id
      - name
      - district
      - kind
      - latitude
      - longitude
      - product_count
      - average_price_index
      - last_reported_at
      title: StoreDetail
...
    TrendsResponse:
      properties:
        period_days:
          type: integer
          title: Period Days
        items:
          items:
            $ref: '#/components/schemas/ProductTrend'
          type: array
          title: Items
      type: object
      required:
      - period_days
      - items
      title: TrendsResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError